+++ dubbo-config/dubbo-config-api/src/main/java/org/apache/dubbo/config/validator/MethodConfigValidator.java
@@ -19,7 +19,7 @@ package org.apache.dubbo.config.validator;
 import org.apache.dubbo.common.extension.Activate;
 import org.apache.dubbo.config.MethodConfig;
 import org.apache.dubbo.config.context.ConfigValidator;
-import org.apache.dubbo.config.util.ConfigValidationUtils;
+import org.apache.dubbo.config.utils.ConfigValidationUtils;
 import org.apache.dubbo.rpc.cluster.LoadBalance;
 
 import static org.apache.dubbo.common.constants.CommonConstants.LOADBALANCE_KEY;
