+++ dubbo-config/dubbo-config-api/src/main/java/org/apache/dubbo/config/validator/ReferenceConfigValidator.java
@@ -21,7 +21,7 @@ import org.apache.dubbo.config.ConsumerConfig;
 import org.apache.dubbo.config.ReferenceConfig;
 import org.apache.dubbo.config.RegistryConfig;
 import org.apache.dubbo.config.context.ConfigValidator;
-import org.apache.dubbo.config.util.ConfigValidationUtils;
+import org.apache.dubbo.config.utils.ConfigValidationUtils;
 import org.apache.dubbo.rpc.InvokerListener;
 
 import java.util.List;
