+++ dubbo-config/dubbo-config-api/src/main/java/org/apache/dubbo/config/validator/ServiceConfigValidator.java
@@ -22,7 +22,7 @@ import org.apache.dubbo.config.ProviderConfig;
 import org.apache.dubbo.config.RegistryConfig;
 import org.apache.dubbo.config.ServiceConfig;
 import org.apache.dubbo.config.context.ConfigValidator;
-import org.apache.dubbo.config.util.ConfigValidationUtils;
+import org.apache.dubbo.config.utils.ConfigValidationUtils;
 import org.apache.dubbo.rpc.ExporterListener;
 
 import java.util.List;
