+++ dubbo-monitor/dubbo-monitor-common/src/main/java/org/apache/dubbo/config/validator/MonitorConfigValidator.java
@@ -21,7 +21,7 @@ import org.apache.dubbo.common.logger.ErrorTypeAwareLogger;
 import org.apache.dubbo.common.logger.LoggerFactory;
 import org.apache.dubbo.config.MonitorConfig;
 import org.apache.dubbo.config.context.ConfigValidator;
-import org.apache.dubbo.config.util.ConfigValidationUtils;
+import org.apache.dubbo.config.utils.ConfigValidationUtils;
 
 
 @Activate
